backend/requirements.txt
numpy==1.26.2
duckduckgo-search==4.1.1

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Google AI & Sheets
google-generativeai==0.3.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
oauth2client==4.1.3

# Async & HTTP
aiohttp==3.9.1
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Utilities
pyyaml==6.0.1

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.23.2